extract Spinner Adapter class into new file CameraSpinAdapter.kt
